Rule changes based on current guide
All the following rules I've changed are based on what I found
in the guide, that were probably being mis-represented in the scss
-lint config.

- import-path-filename-extension: Webpack often needs file extensions to
function correctly, so I think this rule should be turned off.
- declaration-block-max-declarations: we don't actually specify anywhere
in the guide that you cannot have more then 10 declarations. I think
this was mistakenly brought in from another config.
- selector-pseudo-element-colon-notation: saying two colons isn't
specified in the guide so I took it out. Is it required at all?
- value-list-comma-newline-after: this was changed to what I think we
wanted it to be.

Ran the tests and they passed.